requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
face_recognition==1.3.0
opencv-python-headless==4.7.0.72
numpy==1.24.0
pillow==9.3.0
nest_asyncio==1.5.6
python-multipart==0.0.5